메인 프로세스 처리 쓰레드 번위 변경

+++ Trunk/JG/GSNetwork/GSExecuteArgument.h
@@ -11,8 +11,8 @@ public:
 	~GSExecuteArgument()	{}
 
 	GSProactorImpl	*m_ProcactorImpl;
-	BYTE			m_ExecuteType;
-	WORD			m_Id;
+	int				m_ExecuteType;
+	int				m_Id;
 
 };
